@@ -21,8 +21,8 @@
     "url": "git+https://github.com/jay0815/react_from_zero.git"
   },
   "keywords": [
-    "react(v15)",
-    "react-router",
+    "react(v16)",
+    "react-router(v4)",
     "redux",
     "es6"
   ],
@@ -33,7 +33,7 @@
   },
   "homepage": "https://github.com/jay0815/react_from_zero#readme",
   "dependencies": {
-    "antd": "^3.0.0",
+    "antd": "^3.1.3",
     "autoprefixer": "^7.1.6",
     "babel-plugin-import": "^1.6.2",
     "babel-polyfill": "^6.26.0",
@@ -47,7 +47,9 @@
     "es6-promise": "^4.1.1",
     "events": "^1.1.1",
     "express": "^4.15.4",
+    "express-static-gzip": "^0.3.2",
     "file-loader": "^0.11.2",
+    "flux-standard-action": "^2.0.0",
     "history": "^4.7.2",
     "html-webpack-plugin": "^2.30.1",
     "https": "^1.0.0",
@@ -62,8 +64,8 @@
     "prop-types": "^15.6.0",
     "raf": "^3.4.0",
     "rc-queue-anim": "^1.3.1",
-    "react": "^16.1.1",
-    "react-dom": "^16.1.1",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
     "react-hot-loader": "^3.1.1",
     "react-redux": "^5.0.5",
     "react-router-dom": "^4.2.2",
@@ -75,9 +77,9 @@
     "request": "^2.82.0",
     "rucksack-css": "^1.0.2",
     "superagent": "^3.8.0",
-    "webpack": "^3.8.1"
-  },
-  "devDependencies": {
+    "webpack": "^3.8.1",
+    "webpack-dev-middleware": "^1.12.0",
+    "webpack-hot-middleware": "^2.18.2",
     "babel-core": "^6.26.0",
     "babel-istanbul": "^0.12.2",
     "babel-loader": "^7.1.2",
@@ -88,33 +90,35 @@
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-1": "^6.24.1",
     "babel-register": "^6.26.0",
-    "chai": "^4.1.2",
+    "compression-webpack-plugin": "^1.1.3",
     "connect-history-api-fallback": "^1.4.0",
-    "css-loader": "^0.28.7",
     "duplicate-package-checker-webpack-plugin": "^2.0.1",
-    "enzyme": "^3.2.0",
-    "enzyme-adapter-react-16": "^1.1.0",
+    "style-loader": "^0.18.2",
+    "url-loader": "^0.5.9",
+    "rimraf": "^2.6.2",
+    "css-loader": "^0.28.7",
     "extract-text-webpack-plugin": "^3.0.2",
-    "happypack": "^4.0.0",
     "ignore-styles": "^5.0.1",
-    "istanbul": "^1.1.0-alpha.1",
-    "jsdom": "^11.4.0",
+    "happypack": "^4.0.0",
     "less": "^2.7.3",
     "less-loader": "^4.0.5",
-    "mocha": "^4.0.1",
-    "mochawesome": "^2.3.1",
     "npm-run-all": "^4.1.1",
     "postcss-loader": "^2.0.6",
     "postcss-scss": "^1.0.2",
-    "react-test-renderer": "^16.1.1",
     "redux-devtools": "^3.4.0",
-    "rimraf": "^2.6.2",
-    "style-loader": "^0.18.2",
+    "svg-url-loader": "^2.3.0"
+  },
+  "devDependencies": {
+    "chai": "^4.1.2",
+    "enzyme": "^3.2.0",
+    "enzyme-adapter-react-16": "^1.1.0",
+    "istanbul": "^1.1.0-alpha.1",
+    "jsdom": "^11.4.0",
+    "mocha": "^4.0.1",
+    "mochawesome": "^2.3.1",
+    "react-test-renderer": "^16.1.1",
+    "sinon": "^4.1.6",
     "stylelint": "^8.2.0",
-    "stylelint-config-standard": "^17.0.0",
-    "svg-url-loader": "^2.3.0",
-    "url-loader": "^0.5.9",
-    "webpack-dev-middleware": "^1.12.0",
-    "webpack-hot-middleware": "^2.18.2"
+    "stylelint-config-standard": "^17.0.0"
   }
 }
