@@ -1,12 +1,13 @@
 const express = require('express'),
     bodyParser = require('body-parser'),
     cookieParser = require('cookie-parser'),
-    // configObject = require("./config"),
+    // configObject = require('./config'),
     jwt = require('jsonwebtoken'),
     app = express(),
     request = require('request'),
     path = require('path'),
-    log4js = require("log4js"),
+    log4js = require('log4js'),
+    fs = require('fs'),
     EventEmitter = require('events').EventEmitter;
 var jsonParser = bodyParser.json();
 
@@ -15,11 +16,22 @@ var readyPromise = new Promise(resolve => {
   _resolve = resolve
 });
 
+/**
+ * 判断是否有logs文件夹，没有的话自动加上，防止下面日志文件报错
+ */
+if (!fs.existsSync('logs')) {
+  fs.mkdirSync('logs')
+} else {
+  if (!fs.statSync('logs').isDirectory()) {
+    throw('there has a file <<"logs">>, place rename or delete is.')
+  }
+}
+
 log4js.configure({
     appenders: {
         out: { type: 'console' },
-        result: { type: 'dateFile', filename: 'logs/normal.log',layout: {type: 'basic'},"pattern":".yyyy-MM-dd.log", alwaysIncludePattern:true},
-        default: { type: 'dateFile', filename: 'logs/default.log', layout: {type: 'basic'},"pattern":"yyyy-MM-dd.log",alwaysIncludePattern:true},
+        result: { type: 'dateFile', filename: 'logs/normal.log',layout: {type: 'basic'},'pattern':'.yyyy-MM-dd.log', alwaysIncludePattern:true},
+        default: { type: 'dateFile', filename: 'logs/default.log', layout: {type: 'basic'},'pattern':'yyyy-MM-dd.log',alwaysIncludePattern:true},
     },
     categories: {
         default: {
@@ -31,7 +43,7 @@ log4js.configure({
     }
 });
 
-const normalLog = log4js.getLogger("normal");
+const normalLog = log4js.getLogger('normal');
 
 app.use(log4js.connectLogger(normalLog, {level: 'trace', format: ':method :url'}));
 
@@ -43,9 +55,15 @@ const isDeveloping = !isProduction;
 
 app.use(cookieParser());
 
-
 //通过localhost可以访问项目文件夹下的所有文件，等于动态为每个静态文件创建了路由
-app.use(express.static(path.join(__dirname, '/dist')))
+app.use(express.static(path.join(__dirname, '/dist')));
+
+// const publicPath = path.join(__dirname + '/dist');
+// 加入gzip解析 .gz后缀文件
+// if ( isProduction ) {
+//   const expressStaticGzip = require('express-static-gzip')
+//   app.use('/', expressStaticGzip(publicPath))
+// }
 
 app.get(['/*'], function (req, res) {
   res.sendFile(path.join(__dirname, '/dist/index.html'));
@@ -88,7 +106,7 @@ app.post('/login', jsonParser, function (req, res) {
 					let userResult = JSON.parse(body);
 
 					if (userResult.code != 0) {
-						res.json({code: "-1", message: userResult.message});
+						res.json({code: '-1', message: userResult.message});
 						return;
 					}
 
@@ -99,7 +117,7 @@ app.post('/login', jsonParser, function (req, res) {
 
 					return;
 				} else {
-					event.emit('requestError', res, "登录【" + username + "," + pwd + "】", body, response, error);
+					event.emit('requestError', res, '登录【' + username + ',' + pwd + '】', body, response, error);
 					return;
 				}
 			});
