@@ -3,21 +3,21 @@ import types from '../Store/types';
  * demo
  * @param {[type]} param [description]
  */
-export function demoSetState () {
-	return (dispatch, getState) => {
-		dispatch({
-			type: types.SET_APP_STATE
-		});
-	};
+export const demoSetState = () => {
+  return (dispatch, getState) => {
+    dispatch({
+      type: types.SET_APP_STATE
+    });
+  };
 }
-export function change (param) {
-	return (dispatch, getState) => {
-		dispatch({
-			type: types.CHANGE,
-			params: {
-				start: param === 'start' ? true : false,
-				stop: param === 'stop' ? true : false,
-			}
-		});
-	}
+export const change = (param) => {
+  return (dispatch, getState) => {
+    dispatch({
+      type: types.CHANGE,
+      params: {
+        start: param === 'start',
+        stop: param === 'stop',
+      }
+    });
+  }
 }
