@@ -1,11 +1,11 @@
 import keyMirror from 'key-mirror';
 
-
 export default keyMirror({
-	// Home
-	SET_APP_STATE: null,
-	CHANGE: null,
-	// Login
-	SET_LOGIN_INFO: null,
-	CHANGE_LONGIN_STATE: null
+  // Home
+  SET_APP_STATE: null,
+  CHANGE: null,
+  // Login
+  LOGIN: null,
+  SET_LOGIN_INFO: null,
+  CHANGE_LONGIN_STATE: null
 });
