@@ -1,8 +1,16 @@
-export const noop = function noop () {};
 // 抽像 Reducers 生成器
 export function reducerCreators (initialState, actionTypeMapList) {
-	return (state = initialState, action) => {
-		const reducerInstance = typeof actionTypeMapList === 'object' && actionTypeMapList[action.type] ? actionTypeMapList[action.type](state, action) : state;
-		return reducerInstance;
-	};
+  return (state = initialState, action) => {
+    // 校验
+    const reducerInstance = typeof actionTypeMapList === 'object' && actionTypeMapList[action.type] ? actionTypeMapList[action.type](state, action.payload ? action.payload : {}, action.params) : state;
+    return reducerInstance;
+  };
 }
+// reuducer 中抽离 params的版本
+// export function reducerCreators (initialState, actionTypeMapList) {
+//   return (state = initialState, action) => {
+//     // 校验
+//     const reducerInstance = typeof actionTypeMapList === 'object' && actionTypeMapList[action.type] ? actionTypeMapList[action.type](state, action) : state;
+//     return reducerInstance;
+//   };
+// }
