@@ -7,7 +7,7 @@ const path = require('path');
 // import ConsoleLogOnBuildWebpackPlugin from './normal';
 const webpack = require('webpack');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-const precss = require('precss');
+// const precss = require('precss');
 const autoprefixer = require('autoprefixer');
 const rucksackCss = require('rucksack-css');
 const ConsoleLogOnBuildWebpackPlugin = require('./normal');
@@ -15,7 +15,7 @@ const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const theme = require('./theme.js');
 const DuplicatePackageCheckerPlugin = require('duplicate-package-checker-webpack-plugin');
 const HappyPack = require('happypack');
-
+// const CompressionPlugin = require("compression-webpack-plugin");
 const happyThreadPool = HappyPack.ThreadPool({ size: 4 });
 
 // import px2rem from 'postcss-pxtorem';
@@ -24,35 +24,33 @@ const happyThreadPool = HappyPack.ThreadPool({ size: 4 });
 // var autoprefixer = require('autoprefixer');
 // var pxtorem = require('postcss-pxtorem'); //移动端适配使用
 module.exports = {
-	// debug: true, loaders 的 debug 模式将在 webpack 3 或后续版本中取消。
-	entry: {
-		// 文件入口配置
-		index: './src/index',
-		vendor: [
-			'react',
-			'react-dom',
-			'react-router-dom',
-			'redux',
-			// 'redux-logger',
-			// 'redux-saga',
-			'redux-thunk',
-		]
-		// 为了优化，切割代码，提取第三方库（实际上，我们将会引入很多第三方库）
-	},
-	output: {
-		// 文件输出配置
-		path: path.join(__dirname, 'dist'),
-		// 输出目录的配置，模板、样式、脚本、图片等资源的路径配置都相对于它.
-		publicPath: '/',
-		// 模板、样式、脚本、图片等资源对应的server上的路径
-		// filename: 'index.js',
-
-		filename: '[name].[chunkhash].js',
-
-		chunkFilename: '[name].[chunkhash].js',
-		// 命名生成的JS
-	},
-	plugins: [
+  // debug: true, loaders 的 debug 模式将在 webpack 3 或后续版本中取消。
+  entry: {
+  // 文件入口配置
+    index: './src/index',
+    vendor: [
+      'react',
+      'react-dom',
+      'react-router-dom',
+      'redux',
+      // 'redux-logger',
+      // 'redux-saga',
+      'redux-thunk',
+    ]
+    // 为了优化，切割代码，提取第三方库（实际上，我们将会引入很多第三方库）后面会考虑使用cdn
+  },
+  output: {
+    // 文件输出配置
+    path: path.join(__dirname, 'dist'),
+    // 输出目录的配置，模板、样式、脚本、图片等资源的路径配置都相对于它.
+    publicPath: '/',
+    // 模板、样式、脚本、图片等资源对应的server上的路径
+    // filename: 'index.js',
+    filename: '[name].[chunkhash].js',
+    chunkFilename: '[name].[chunkhash].js',
+    // 命名生成的JS
+  },
+  plugins: [
 		// new webpack.optimize.OccurrenceOrderPlugin(), webpack 2 or 3 及 更高版本中 OccurrenceOrderPlugin 被默认加载
 		new HtmlWebpackPlugin({
 			template: './Template/index.html',
@@ -112,11 +110,20 @@ module.exports = {
 			debug: false
 		}),
 		new DuplicatePackageCheckerPlugin(),
+    new ConsoleLogOnBuildWebpackPlugin(),
 		new HappyPack({
 	      id: 'js',
 	      threadPool: happyThreadPool,
 	      loaders: [ 'babel-loader' ]
-	  	})
+    })
+    // 文件二次压缩 ，效果要好于cdn
+    // new CompressionPlugin({
+    //   test: /\.(js|css)$/,
+    //   asset: '[path].gz[query]',
+    //   algorithm: 'gzip',
+    //   threshold: 10240,
+    //   minRatio: 0.8
+    // })
 	],
 	module: {
 		rules: [
@@ -184,7 +191,6 @@ module.exports = {
 			},
 			{
 				test: /\.(svg).*$/i,
-				// include: svgSpriteDirs, // 把 svgDirs 路径下的所有 svg 文件交给 svg-sprite-loader 插件处理
 				use: [{
 					loader: 'url-loader?limit=1',
 					options: {
